Check_Valid_Report {rsell}	R Documentation

Check the Validity of the Input Poshmark Sales Activity File

Description

This function Checks the Validity of the Input Poshmark Sales Activity File.

Usage

Check_Valid_Report(t_datapath)

Arguments

t_datapath
The file to be read.

Value

valid_report, valid_user Logical values in array.

[Package rsell version 1.0.0 Index]