Update_Quantity_Add {rsell}	R Documentation

Update the Inventory Items to be Added

Description

This function creates additional records for inventory add based on the item quantity.

Usage

Update_Quantity_Add(inv_record, item_qty)

Arguments

inv_record
The Inventory Record to be added.

item_qty
The number of items to be added.

Details

Requires:

Value

inv_record The updated data frame with all inventory records to be added.

[Package rsell version 1.0.0 Index]